Add files via upload
This research delves into the enigma of Venezuela's remarkable GDP surge from -8% in 2003 to an astonishing 18% in 2004. Amidst a backdrop of economic struggles, this study investigates the pivotal role played by a governmental exchange rate policy. During this critical period, the Venezuelan Ministry of Finance implemented a radical policy shift, wresting control of the exchange rate market. This transformed the exchange rate dynamics between the Venezuelan Bolivar (VEF) and the US dollar (USD), imposing a fixed rate of 1,853 VEF to 1,600 VEF per USD. The resulting forced appreciation of the Bolivar disrupted market mechanisms, extinguishing vital informational signals for consumers, producers, and investors, while simultaneously nurturing an atmosphere conducive to corruption.

Employing a comprehensive analysis of historical data encompassing GDP, export, and import growth, as well as the corruption index from the Heritage Foundation, this research unveils the driving forces behind this perplexing GDP surge. The findings conclusively reveal that the meteoric rise in GDP was exclusively orchestrated by the manipulation of the exchange rate and not attributed to increased oil production or alternative sources of income. Moreover, this study underscores the detrimental consequences of this policy, as corruption surged by 11% during this period and continued its ascent thereafter. In essence, this research demystifies the illusion of Venezuela's GDP growth, shedding light on the dark underbelly of a policy that ultimately proved detrimental to the nation's economic stability.